openai-agent-sdk/agent-as-tool/pyproject.toml
[project]
name = "agent-as-tool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.5.5",
    "openai-agents>=0.0.12",
    "python-dotenv>=1.1.0",
]
